\name{Rjungle-package}
\alias{Rjungle-package}
\alias{Rjungle}
\docType{package}
\title{
Rjungle
}
\description{
Performs RandomJungle from within R.
}
\details{
\tabular{ll}{
Package: \tab Rjungle\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2010-04-05\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{
Daniel Schwarz with minor modifications by Andreas Bender
}
\references{
www.randomjungle.org
}
\keyword{ package }
\seealso{
\code{\link{rjungle}}, \code{\link{importance}}
}
\examples{

}
