\name{rjungleR-package}
\alias{rjungleR-package}
\alias{rjungle}
\docType{package}
\title{
rjungleR
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab Rjungle\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2010-04-05\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{

}
\references{
}
\keyword{ package }
\seealso{

\code{\link{rjungle}}, \code{\link{importance}}
}
\examples{

}
